# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Antyakshari Event Registration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegistrationPaymentStatus = typeof RegistrationPaymentStatus[keyof typeof RegistrationPaymentStatus];


export const RegistrationPaymentStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Registration {
  id: number;
  bookingId: string;
  name: string;
  email: string;
  phone: string;
  totalPasses: number;
  /** @nullable */
  paymentScreenshot?: string | null;
  paymentStatus: RegistrationPaymentStatus;
  createdAt: string;
}

export type RegistrationWithParticipantsPaymentStatus = typeof RegistrationWithParticipantsPaymentStatus[keyof typeof RegistrationWithParticipantsPaymentStatus];


export const RegistrationWithParticipantsPaymentStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Participant {
  id: number;
  registrationId: number;
  participantName: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  collegeOrCompany?: string | null;
  /** @nullable */
  passId?: string | null;
  /** @nullable */
  qrToken?: string | null;
  isUsed?: boolean;
}

export interface RegistrationWithParticipants {
  id: number;
  bookingId: string;
  name: string;
  email: string;
  phone: string;
  totalPasses: number;
  /** @nullable */
  paymentScreenshot?: string | null;
  paymentStatus: RegistrationWithParticipantsPaymentStatus;
  createdAt: string;
  participants: Participant[];
}

export interface ParticipantInput {
  /** @minLength 2 */
  participantName: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  collegeOrCompany?: string | null;
}

export interface RegistrationInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 10 */
  phone: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  totalPasses: number;
  participants: ParticipantInput[];
}

export interface PaymentScreenshotInput {
  screenshotUrl: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminAuthResult {
  success: boolean;
  token: string;
}

export interface AdminStats {
  totalRegistrations: number;
  pendingRegistrations: number;
  approvedRegistrations: number;
  rejectedRegistrations: number;
  totalPasses: number;
  totalPassesIssued: number;
}

export type PassValidationResultStatus = typeof PassValidationResultStatus[keyof typeof PassValidationResultStatus];


export const PassValidationResultStatus = {
  valid: 'valid',
  used: 'used',
  invalid: 'invalid',
} as const;

export interface PassValidationResult {
  status: PassValidationResultStatus;
  /** @nullable */
  passId: string | null;
  /** @nullable */
  participantName: string | null;
  eventName: string;
  eventDate: string;
  isUsed?: boolean;
}

export type ListRegistrationsParams = {
status?: ListRegistrationsStatus;
};

export type ListRegistrationsStatus = typeof ListRegistrationsStatus[keyof typeof ListRegistrationsStatus];


export const ListRegistrationsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;