# lib/api-zod/src/generated/types/participant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Antyakshari Event Registration API
 * OpenAPI spec version: 0.1.0
 */

export interface Participant {
  id: number;
  registrationId: number;
  participantName: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  collegeOrCompany?: string | null;
  /** @nullable */
  passId?: string | null;
  /** @nullable */
  qrToken?: string | null;
  isUsed?: boolean;
}