# lib/api-zod/src/generated/types/participantInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Antyakshari Event Registration API
 * OpenAPI spec version: 0.1.0
 */

export interface ParticipantInput {
  /** @minLength 2 */
  participantName: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  collegeOrCompany?: string | null;
}