# lib/api-zod/src/generated/types/passValidationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Antyakshari Event Registration API
 * OpenAPI spec version: 0.1.0
 */
import type { PassValidationResultStatus } from './passValidationResultStatus';

export interface PassValidationResult {
  status: PassValidationResultStatus;
  /** @nullable */
  passId: string | null;
  /** @nullable */
  participantName: string | null;
  eventName: string;
  eventDate: string;
  isUsed?: boolean;
}