# lib/api-zod/src/generated/types/paymentScreenshotInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Antyakshari Event Registration API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentScreenshotInput {
  screenshotUrl: string;
}