# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Antyakshari Event Registration API
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantInput } from './participantInput';

export interface RegistrationInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 10 */
  phone: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  totalPasses: number;
  participants: ParticipantInput[];
}